feat: Add useAdminAllUsers hook for fetching all users in the admin dashboard
This commit adds a new hook, useAdminAllUsers, which is responsible for fetching all users in the admin dashboard. It uses the useQuery function from the @tanstack/react-query library to make the API request to retrieve the user data. The hook returns the user data, loading state, error state, and a refetch function for refreshing the data. This hook will be used in the UsersDashboard component to display the list of users and perform actions like deleting a user.