feat: create product dashboard route for show all products so admin can see all product and manage it
feat: Add OrdersDashboard component and update admin routes

This commit adds the OrdersDashboard component to the project, which will be used in the admin dashboard to display orders. Additionally, the admin routes have been updated to include the new OrdersDashboard component. This allows users with admin privileges to access and manage orders in the dashboard.