feat: Add ProductsDashboardAddForm component for adding new products
This commit adds the ProductsDashboardAddForm component to the ProductsDashboard. The form allows users to add new products to the dashboard. When the "Add Product" button is clicked, the form is displayed, and users can enter the necessary information for the new product. After submitting the form, the product is added to the database and the product list is refreshed to display the new product. This feature enhances the functionality of the admin dashboard by enabling easy addition of products.